prowloader/releaseloader: deduplicate label queries
